2.1.9  - Fixed a warning in KWSRandomName.m; Now when I push to Cocoapods I do so w/o any warnings;

+++ Pod/Classes/Services/KWS/KWSRandomName.m
@@ -15,7 +15,7 @@
 @implementation KWSRandomName
 
 - (NSString*) getEndpoint {
-    return [NSString stringWithFormat:@"v2/apps/%ld/random-display-name", appId];
+    return [NSString stringWithFormat:@"v2/apps/%ld/random-display-name", (long)appId];
 }
 
 - (KWS_HTTP_METHOD) getMethod {
